storage/model/requirements.txt
ultralytics
easyocr
python-dotenv
